# py/gc: Skip marking inside buffers of str, bytes, bytearray.