These Python objects all contain pointers onto (often large) regions of the Python heap that we know won't contain any pointers to the heap themselves, so we can skip scanning them in the "mark" pass.

This is a particularly large performance boost on platforms with large slow RAM buses (i.e. ESP32 series with PSRAM).

## Benchmarks

<details><summary>Click to see some Python code that demonstrates best-case improvements</summary>

```py
import gc
import time

def time_gc():
    d = gc.mem_free()
    s = time.ticks_us()
    gc.collect()
    s = time.ticks_us() - s
    f = gc.mem_free()
    # Note: If MICROPY_GC_SPLIT_HEAP_AUTO is added & enabled, this will
    # sometimes report negative numbers of freed bytes due to reclaiming heap
    # areas(!)
    d = gc.mem_free() - d
    print(f'Collect took {s}us freed {d} bytes, now {f} bytes free')

time_gc()

a = b'x' * 100_000
b = b'y' * 100_000
c = b'z' * 100_000

time_gc()

d = bytearray(100_000)

time_gc()

del a
del b
del c
del d

time_gc()
```

</details>

Measured on an ESP32 ECO1 with 4MB PSRAM and `GENERIC_SPIRAM` board config.

Current master:

```
Collect took 611us freed 576 bytes, now 2180736 bytes free
Collect took 35605us freed 96 bytes, now 1880528 bytes free
Collect took 47925us freed 176 bytes, now 1780512 bytes free
Collect took 8503us freed 400272 bytes, now 2180608 bytes free
```

This branch:

```
Collect took 603us freed 576 bytes, now 2180736 bytes free
Collect took 4500us freed 80 bytes, now 1880512 bytes free
Collect took 5819us freed 176 bytes, now 1780496 bytes free
Collect took 8478us freed 400288 bytes, now 2180608 bytes free
```

When running the middle two GC passes, marking has to scan ~300KB and ~400KB fewer bytes and saves ~31ms and ~42ms of time, respectively.

## Notes on implementation

The implementation is very simple, it compares a possible `mp_obj_base_t->type` pointer at the head to the three types where this applies (`mp_type_str`, `mp_type_bytes`, `mp_type_bytearray`).

If someone allocates a C-style struct in the Python heap where the first pointer happens to point to one of these three `mp_type_nnn` entities, use-after-free corruption will occur... However, I don't see anything similar to this in the MicroPython codebase at present.

It's tempting to try and implement something more general (for example, add a `MP_TYPE_FLAG_GC_SKIP` bit to `mp_obj_type_t` flags or even implement a way for Python types to describe which of their pointers can be skipped.)  This could allow optimisation of other types like StringIO/BytesIO. However it seems hard to safely resolve whether or not a heap entry holds a Python object without adding storage overhead or linker tricks.

---

This work was funded through GitHub Sponsors.