@projectgus 2023-08-08
> It's tempting to try and implement something more general [..] However it seems hard to safely resolve whether or not a heap entry holds a Python object without adding storage overhead or linker tricks.

For the sake of talking this through a little more, the best way that I could think of to implement this in a general fashion:

* Use port linker scripts to place all `mp_obj_type_t` structs into one region of .rodata, with the linker providing markers for region start & end.
* Compare each possible base obj type pointer to this region. If it's inside the limits of the "type struct" region and aligns with one of the struct entries, assume it's a Python object of this type.
* Then safe to deference the type pointer and read metadata about how to handle it in the mark pass (either a flag bit, or for some types it may be useful for it to say "skip scanning this particular pointer inside this object").

The implementation in this PR is a *lot* simpler, though, for a similar result!