@projectgus 2023-08-08
Closing this one out as @dpgeorge rightly points out it's not safe enough for the case where a C structure may have a first entry that happens to line up with the address of one of these types.

@jimmo has had the excellent idea that, instead, the finalizer bit in the FTB could be converted to "is python object" bit, allowing the GC to reliably know if it's looking at a Python object, or not.