@dpgeorge 2023-08-08
Unfortunately I don't think this is precise enough.  Consider the case:
```c
struct _uart_t {
    uint32_t counter;
    uint8_t *rx_buf;
};
```
It could be that `uart->counter` counts up to `&mp_type_str`, in which case the GC will no longer trace this struct and will free the `rx_buf` entry.