@dpgeorge 2023-08-08
Garbage collectors are hard!